backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
markdown==3.6
pygments==2.17.2
python-slugify==8.0.4
aiofiles==23.2.1
pydantic-settings==2.2.1
httpx==0.27.0
